+++ svd-encoder/src/peripheral.rs
@@ -1,17 +1,22 @@
-use super::{new_node, Element, ElementMerge, Encode, EncodeChildren, EncodeError, XMLNode};
+use super::{
+    new_node, Config, Element, ElementMerge, Encode, EncodeChildren, EncodeError, XMLNode,
+};
 
-use crate::svd::{Interrupt, Peripheral, PeripheralInfo};
+use crate::{
+    config::{change_case, format_number},
+    svd::{Peripheral, PeripheralInfo},
+};
 
 impl Encode for Peripheral {
     type Error = EncodeError;
 
-    fn encode(&self) -> Result<Element, EncodeError> {
+    fn encode_with_config(&self, config: &Config) -> Result<Element, EncodeError> {
         match self {
-            Self::Single(info) => info.encode(),
+            Self::Single(info) => info.encode_with_config(config),
             Self::Array(info, array_info) => {
                 let mut base = Element::new("peripheral");
-                base.merge(&array_info.encode()?);
-                base.merge(&info.encode()?);
+                base.merge(&array_info.encode_with_config(config)?);
+                base.merge(&info.encode_with_config(config)?);
                 Ok(base)
             }
         }
@@ -21,9 +26,12 @@ impl Encode for Peripheral {
 impl Encode for PeripheralInfo {
     type Error = EncodeError;
 
-    fn encode(&self) -> Result<Element, EncodeError> {
+    fn encode_with_config(&self, config: &Config) -> Result<Element, EncodeError> {
         let mut elem = Element::new("peripheral");
-        elem.children.push(new_node("name", self.name.clone()));
+        elem.children.push(new_node(
+            "name",
+            change_case(&self.name, config.peripheral_name),
+        ));
 
         if let Some(v) = &self.display_name {
             elem.children.push(new_node("displayName", v.to_string()));
@@ -38,8 +46,10 @@ impl Encode for PeripheralInfo {
         }
 
         if let Some(v) = &self.alternate_peripheral {
-            elem.children
-                .push(new_node("alternatePeripheral", v.to_string()));
+            elem.children.push(new_node(
+                "alternatePeripheral",
+                change_case(v, config.peripheral_name),
+            ));
         }
 
         if let Some(v) = &self.group_name {
@@ -47,39 +57,55 @@ impl Encode for PeripheralInfo {
         }
 
         if let Some(v) = &self.prepend_to_name {
-            elem.children.push(new_node("prependToName", v.to_string()));
+            elem.children.push(new_node(
+                "prependToName",
+                change_case(v, config.peripheral_name),
+            ));
         }
 
         if let Some(v) = &self.append_to_name {
-            elem.children.push(new_node("appendToName", v.to_string()));
+            elem.children.push(new_node(
+                "appendToName",
+                change_case(v, config.peripheral_name),
+            ));
         }
 
         if let Some(v) = &self.header_struct_name {
-            elem.children
-                .push(new_node("headerStructName", v.to_string()));
+            elem.children.push(new_node(
+                "headerStructName",
+                change_case(v, config.peripheral_name),
+            ));
         }
 
         elem.children.push(new_node(
             "baseAddress",
-            format!("0x{:.08X}", self.base_address),
+            format_number(self.base_address, config.peripheral_base_address),
         ));
 
-        elem.children
-            .extend(self.default_register_properties.encode()?);
+        elem.children.extend(
+            self.default_register_properties
+                .encode_with_config(config)?,
+        );
 
         if let Some(v) = &self.address_block {
             for ab in v {
-                elem.children.push(ab.encode_node()?);
+                elem.children.push(ab.encode_node_with_config(config)?);
             }
         }
 
-        let interrupts: Result<Vec<_>, _> =
-            self.interrupt.iter().map(Interrupt::encode_node).collect();
+        let interrupts: Result<Vec<_>, _> = self
+            .interrupt
+            .iter()
+            .map(|interrupt| interrupt.encode_node_with_config(config))
+            .collect();
 
         elem.children.append(&mut interrupts?);
 
         if let Some(v) = &self.registers {
-            let children: Result<Vec<_>, _> = v.iter().map(|e| e.encode_node()).collect();
+            let children: Result<Vec<_>, _> = v
+                .iter()
+                .map(|e| e.encode_node_with_config(config))
+                .collect();
 
             elem.children.push({
                 let mut e = Element::new("registers");
@@ -89,8 +115,10 @@ impl Encode for PeripheralInfo {
         }
 
         if let Some(v) = &self.derived_from {
-            elem.attributes
-                .insert(String::from("derivedFrom"), v.to_string());
+            elem.attributes.insert(
+                String::from("derivedFrom"),
+                change_case(v, config.peripheral_name),
+            );
         }
 
         Ok(elem)