+++ tests/src/lib.rs
@@ -10,7 +10,7 @@ use xmltree::Element;
 mod bad_svd;
 
 /// Generic test helper function
-/// Takes an array of (item, xml) pairs where the item implements
+/// Takes an array of (item, xml, xml) pairs where the item implements
 /// Parse and Encode and tests object encoding and decoding
 pub fn run_test<
     T: Parse<Error = SVDErrorAt, Object = T, Config = Config>
@@ -18,13 +18,16 @@ pub fn run_test<
         + core::fmt::Debug
         + PartialEq,
 >(
-    tests: &[(T, &str)],
+    // tests: [(Object, input string, output string)],
+    tests: &[(T, &str, &str)],
+    parser_config: Option<Config>,
+    encoder_config: Option<svd_encoder::Config>,
 ) {
     for t in tests {
         let rotree = Document::parse(t.1).unwrap();
         let elem = T::parse(
             &rotree.root().first_element_child().unwrap(),
-            &Config::default(),
+            &parser_config.unwrap_or(Config::default()),
         )
         .unwrap();
         assert_eq!(
@@ -32,11 +35,13 @@ pub fn run_test<
             "Error parsing xml` (mismatch between parsed and expected)"
         );
 
-        let tree1 = Element::parse(t.1.as_bytes()).unwrap();
-        let tree2 = elem.encode().unwrap();
+        let tree1 = Element::parse(t.2.as_bytes()).unwrap();
+        let tree2 = elem
+            .encode_with_config(&encoder_config.unwrap_or(svd_encoder::Config::default()))
+            .unwrap();
         assert_eq!(
             tree1, tree2,
-            "Error encoding xml (mismatch between encoded and original)"
+            "Error encoding xml (mismatch between encoded and expected)"
         );
     }
 }